requirements.txt
pillow==9.4.0
opencv-python==4.8.0.76
numpy==1.23.5
